Implemented initial error handling, changed input loop logic
The input loop has been changed to send the server a message indicating the client has disconnected, and entering messages with the length of zero prints to the client that messages must have a length above zero. Currently this stops execution for the client, will likely revisit this to allow for additional inputs after a zero length message.